feat(mobile): bottom sheets через expo-router formSheet (closes #97)
Мигрировал 6 из 7 bottom-sheets на нативный formSheet через
expo-router (без сторонних пакетов): help, students-filters,
generate-codes, intervals-editor, override-editor.

Каждый sheet — отдельный screen в app/, зарегистрирован flat в
root Stack с presentation: formSheet и sheetAllowedDetents:
fitToContents (sheet сам подгоняет высоту под контент).

DiagnosticsFiltersSheet оставлен на старом Modal-паттерне с
кастомными Animated-анимациями (fade backdrop + slide sheet) и
drag-to-dismiss через Gesture.Pan — слишком много контента для
fitToContents на SDK 54 (где flex: 1 в formSheet не работает).

Параметры между parent/sheet передаются через URL search params;
для editor-sheets мутации перенесены внутрь самого sheet'а
(читают server-state из react-query кеша по shared queryKey).

GestureHandlerRootView добавлен в root layout — нужен для
react-native-gesture-handler (используется gesture-handler'ом
expo-router и нашим drag-жестом в DiagnosticsFiltersSheet).

Удалено: 6 старых Sheet-компонентов (-1953 строки).
