+++ packages/psychologist-mobapp/app/(screens)/office-hours.tsx
@@ -1,7 +1,7 @@
-import { useMemo, useState } from "react";
+import { useMemo } from "react";
 import { Pressable, ScrollView, StyleSheet, View } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { Stack } from "expo-router";
+import { Stack, useRouter } from "expo-router";
 import { Ionicons } from "@expo/vector-icons";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Text, Body } from "../../components/ui";
@@ -18,8 +18,6 @@ import type {
   OfficeHoursOverrideEntry,
   OfficeHoursResolved,
 } from "@tirek/shared";
-import { IntervalsEditorSheet } from "../../components/office-hours/IntervalsEditorSheet";
-import { OverrideEditorSheet } from "../../components/office-hours/OverrideEditorSheet";
 
 const DOW_LABELS: Record<number, string> = {
   1: "Пн",
@@ -60,6 +58,7 @@ function formatHumanDate(iso: string): string {
 export default function OfficeHoursScreen() {
   const c = useThemeColors();
   const qc = useQueryClient();
+  const router = useRouter();
   const user = useAuthStore((s) => s.user);
   const psychologistId = user?.id ?? "";
 
@@ -98,49 +97,6 @@ export default function OfficeHoursScreen() {
   const sortedOverrides = useMemo(() => {
     return [...(overrides ?? [])].sort((a, b) => a.date.localeCompare(b.date));
   }, [overrides]);
-
-  const [editingDow, setEditingDow] = useState<OfficeHoursDayOfWeek | null>(null);
-  const [overrideMode, setOverrideMode] = useState<
-    | { kind: "create" }
-    | { kind: "edit"; date: string; intervals: OfficeHoursInterval[]; notes: string | null }
-    | null
-  >(null);
-
-  const editingTemplateRow = editingDow != null ? templateByDow.get(editingDow) : null;
-
-  const upsertTemplateMut = useMutation({
-    mutationFn: ({
-      dayOfWeek,
-      intervals,
-      notes,
-    }: {
-      dayOfWeek: OfficeHoursDayOfWeek;
-      intervals: OfficeHoursInterval[];
-      notes: string | null;
-    }) => officeHoursApi.upsertTemplateDay(dayOfWeek, intervals, notes),
-    onSuccess: () => {
-      qc.invalidateQueries({ queryKey: ["office-hours-template"] });
-      qc.invalidateQueries({ queryKey: ["office-hours-resolve"] });
-      setEditingDow(null);
-    },
-  });
-
-  const upsertOverrideMut = useMutation({
-    mutationFn: ({
-      date,
-      intervals,
-      notes,
-    }: {
-      date: string;
-      intervals: OfficeHoursInterval[];
-      notes: string | null;
-    }) => officeHoursApi.upsertOverrideDay(date, intervals, notes),
-    onSuccess: () => {
-      qc.invalidateQueries({ queryKey: ["office-hours-overrides"] });
-      qc.invalidateQueries({ queryKey: ["office-hours-resolve"] });
-      setOverrideMode(null);
-    },
-  });
 
   const deleteOverrideMut = useMutation({
     mutationFn: (date: string) => officeHoursApi.deleteOverrideDay(date),
@@ -184,7 +140,10 @@ export default function OfficeHoursScreen() {
                   key={dow}
                   onPress={() => {
                     hapticLight();
-                    setEditingDow(dow);
+                    router.push({
+                      pathname: "/intervals-editor",
+                      params: { dayOfWeek: String(dow) },
+                    });
                   }}
                   style={({ pressed }) => [
                     styles.dayCell,
@@ -255,7 +214,7 @@ export default function OfficeHoursScreen() {
           <Pressable
             onPress={() => {
               hapticLight();
-              setOverrideMode({ kind: "create" });
+              router.push("/override-editor");
             }}
             style={({ pressed }) => [
               styles.addBtn,
@@ -315,11 +274,9 @@ export default function OfficeHoursScreen() {
                     style={{ flex: 1 }}
                     onPress={() => {
                       hapticLight();
-                      setOverrideMode({
-                        kind: "edit",
-                        date: ov.date,
-                        intervals: ov.intervals,
-                        notes: ov.notes,
+                      router.push({
+                        pathname: "/override-editor",
+                        params: { date: ov.date },
                       });
                     }}
                   >
@@ -365,33 +322,6 @@ export default function OfficeHoursScreen() {
         )}
       </ScrollView>
 
-      {editingDow != null ? (
-        <IntervalsEditorSheet
-          open={editingDow != null}
-          title={`Шаблон: ${DOW_LABELS[editingDow]}`}
-          initialIntervals={editingTemplateRow?.intervals ?? []}
-          initialNotes={editingTemplateRow?.notes ?? null}
-          saving={upsertTemplateMut.isPending}
-          onClose={() => setEditingDow(null)}
-          onSave={(intervals, notes) => {
-            upsertTemplateMut.mutate({ dayOfWeek: editingDow, intervals, notes });
-          }}
-        />
-      ) : null}
-
-      {overrideMode != null ? (
-        <OverrideEditorSheet
-          open
-          fixedDate={overrideMode.kind === "edit" ? overrideMode.date : undefined}
-          initialIntervals={overrideMode.kind === "edit" ? overrideMode.intervals : []}
-          initialNotes={overrideMode.kind === "edit" ? overrideMode.notes : null}
-          saving={upsertOverrideMut.isPending}
-          onClose={() => setOverrideMode(null)}
-          onSave={(date, intervals, notes) => {
-            upsertOverrideMut.mutate({ date, intervals, notes });
-          }}
-        />
-      ) : null}
     </SafeAreaView>
   );
 }
