+++ packages/psychologist-mobapp/app/_layout.tsx
@@ -3,6 +3,7 @@ import { PersistQueryClientProvider } from "@tanstack/react-query-persist-client
 import { Stack } from "expo-router";
 import { StatusBar } from "react-native";
 import { SafeAreaProvider } from "react-native-safe-area-context";
+import { GestureHandlerRootView } from "react-native-gesture-handler";
 import * as SplashScreen from "expo-splash-screen";
 import {
   useFonts,
@@ -24,6 +25,13 @@ import { useNotifications } from "../lib/notifications";
 initSentry();
 SplashScreen.preventAutoHideAsync().catch(() => {});
 
+const SHEET_OPTIONS = {
+  presentation: "formSheet",
+  sheetAllowedDetents: "fitToContents",
+  sheetGrabberVisible: true,
+  sheetCornerRadius: 24,
+} as const;
+
 function ThemedApp() {
   useNotifications();
 
@@ -36,7 +44,13 @@ function ThemedApp() {
       <StatusBar barStyle="dark-content" />
       <NetworkStatus />
       <OnlineManager />
-      <Stack screenOptions={{ headerShown: false, animation: "fade" }} />
+      <Stack screenOptions={{ headerShown: false, animation: "fade" }}>
+        <Stack.Screen name="generate-codes" options={SHEET_OPTIONS} />
+        <Stack.Screen name="help" options={SHEET_OPTIONS} />
+        <Stack.Screen name="students-filters" options={SHEET_OPTIONS} />
+        <Stack.Screen name="intervals-editor" options={SHEET_OPTIONS} />
+        <Stack.Screen name="override-editor" options={SHEET_OPTIONS} />
+      </Stack>
     </>
   );
 }
@@ -58,20 +72,22 @@ function RootLayout() {
   if (!fontsLoaded) return null;
 
   return (
-    <PersistQueryClientProvider
-      client={queryClient}
-      persistOptions={{ persister: asyncStoragePersister }}
-    >
-      <LanguageProvider>
-        <ThemeProvider>
-          <SafeAreaProvider>
-            <ErrorBoundary>
-              <ThemedApp />
-            </ErrorBoundary>
-          </SafeAreaProvider>
-        </ThemeProvider>
-      </LanguageProvider>
-    </PersistQueryClientProvider>
+    <GestureHandlerRootView style={{ flex: 1 }}>
+      <PersistQueryClientProvider
+        client={queryClient}
+        persistOptions={{ persister: asyncStoragePersister }}
+      >
+        <LanguageProvider>
+          <ThemeProvider>
+            <SafeAreaProvider>
+              <ErrorBoundary>
+                <ThemedApp />
+              </ErrorBoundary>
+            </SafeAreaProvider>
+          </ThemeProvider>
+        </LanguageProvider>
+      </PersistQueryClientProvider>
+    </GestureHandlerRootView>
   );
 }
